CASMPET-7538 1.3.0 cray-etcd-base charts don't work with PSP
The 1.3.x cray-etcd-base chart pulls in a newer version of the
upstream bitnami etcd chart, which includes the addition of
    containerSecurityContext:
      seccompProfile:
        type: "RuntimeDefault"
in several places.  On Kubernetes 1.24, PSP does not allow
"seccompProfile" to be set, whereas with PSS it is allowed
to be set to a few values, including "RuntimeDefault".  So
on CSM 1.7 with PSS implemented through kyverno, this setting
is allowed, but on CSM 1.6 with PSP it is not allowed.

With the "seccompProfile" section conditionally removed when
being deployed with PSP, the chart can be used on both CSM 1.7
and CSM 1.6.